Simple Lister of Applications
We begin with exploration.  The AT-SPI layer is unknown, so we want
some sort of probe to increase our understanding.  We go for a program
to list the applications on the Desktop.  There is no Ruby binding
for AT-SPI, so we use C to start with.

There is not much point in trying to write an automated test yet.  We
have no infrastructure (we're trying to build it!), but more
importantly we don't know what we expect to see.  This doesn't
invalidate TDD -- we know the behaviour we are aiming for (a list of
applications), but we don't know sufficient detail (what counts as an
application?) to write an automated test (yet).  (It does mean that
we don't know what 'done' means, so we have to be careful to do an
appropriate amount of work.)

As we make the code work, we discover several things that can go wrong.
    Assistive Technologies may not be enabled.
    There are no Desktop children until you logon again.
    Resources may not be freed.
    There are NULL children of the Desktop.
For the moment this knowledge is pushed into defensive code.  Later it
should be made explicit with automated tests.

From a broader perspective, we need to be aware that this is the
AT-SPI layer's view of what an Application is.  There will be other
views, and we need to make sure that we end up with a view which is of
maximum use to the end user.