Ruby version of Lister
We now have a Ruby object which performs a similar function to the
previous C code.  It is not a sensible design of object, and it is not
properly supported by tests -- we are still exploring.

Also, the few tests we have will fail as soon as anything in the
environment changes.