Report Frame titles, rather than process titles.
The names we were reporting for the applications were retrieved using
Accessible_getName() on the Accessibles which return true from
Accessible_isApplication() and are direct descendents of the Desktop.
This turns out to be the name of the process (as reported by "ps x").

We actually want the name of the Frame which is the child of
these Applications.

So we currently assume that the hierarchy is
    Desktop -> Application -> Frame
(ie all Applications are direct children of the Desktop, and all
Frames are direct children of Applications)

This matches what we have seen so far, but we must remember that
experience is necessarily limited, and may thus be misleading.