Search for matching Application
We are not producing a general-purpose driver for the AT-SPI layer.
This is a testing tool, so we want to find a matching Frame.
So we launch a simple application and test that we can find it.

We also test that we correctly fail to find one that hasn't been
launched.

After we launch the test app, we sleep for half a second to let it
register itself.  This is completely bogus.  The testing tool needs to
handle these sort of issues.  I'll start dealing with that next.