Find application which initializes concurrently with Driver
If the application has already initialized by the time the Application
Driver is initialized, we will find it as a child of the Desktop.  If
we do not find it, a listener is registered which will be notified
when the application does initialize.

What is missing at the moment is a timeout on waiting for the
application to appear.  So if we try to create an Application Driver
for an app which never starts, the creation call will never return.